--- a/packages/replay-core/src/__benchmarks__/replay-core.benchmark.ts
+++ b/packages/replay-core/src/__benchmarks__/replay-core.benchmark.ts
@@ -103,8 +103,8 @@ export function runInputSprite(arg: SuiteArg) {
     init() {
       return { x: 0 };
     },
-    loop({ state, device }) {
-      if (device.inputs.clicked) {
+    loop({ state, getInputs }) {
+      if (getInputs().clicked) {
         return { ...state, x: state.x + 1 };
       }
       return state;