--- a/packages/replay-core/src/__benchmarks__/utils.ts
+++ b/packages/replay-core/src/__benchmarks__/utils.ts
@@ -20,7 +20,7 @@ export function runGame(
 ) {
   const platform = getBenchmarkPlatform();
 
-  const { getNextFrameTextures } = replayCore(
+  const { runNextFrame } = replayCore(
     platform,
     getNativeSpriteSettings(),
     Game(gameProps)
@@ -32,7 +32,7 @@ export function runGame(
   // Run 1000 frames
   for (let i = 1; i <= 1000; i++) {
     arg.onFrameStart(i);
-    getNextFrameTextures(i * oneFrameMs, resetInputs);
+    runNextFrame(i * oneFrameMs, resetInputs);
     arg.onFrameEnd(i);
   }
 }
@@ -136,6 +136,12 @@ function getBenchmarkPlatform() {
           inputs: { ...inputs, x: local.x, y: local.y },
         };
       };
+    },
+    render: {
+      newFrame: () => null,
+      startRenderSprite: () => null,
+      endRenderSprite: () => null,
+      renderTexture: () => null,
     },
   };