--- a/packages/replay-starter-ts/src/index.ts
+++ b/packages/replay-starter-ts/src/index.ts
@@ -52,10 +52,10 @@ export const Game = makeSprite<GameProps, GameState, WebInputs | iOSInputs>({
     };
   },
 
-  loop({ state, device }) {
+  loop({ state, device, getInputs }) {
     if (!state.loaded) return state;
 
-    const { pointer } = device.inputs;
+    const { pointer } = getInputs();
     const { posX, posY } = state;
     let { targetX, targetY } = state;