Some improvements to memory usage and speed of replay-core. Note this has a **breaking API change**:

- `inputs` has been removed from `device` and replaced with `getInputs` in sprite methods.

Benchmarks results on my machine (node v12.20.1) show ~50% drop in memory usage per frame (less GC) and ~40% speedup:

```
Simple texture
new x 1,358 ops/sec ±1.25% (88 runs sampled) 1016 B / frame
previous x 1,084 ops/sec ±1.05% (89 runs sampled) 2256 B / frame
Fastest is new
Lowest memory is new

Nested Sprite
new x 791 ops/sec ±1.42% (87 runs sampled) 1840 B / frame
previous x 554 ops/sec ±2.50% (86 runs sampled) 4376 B / frame
Fastest is new
Lowest memory is new

Pure Sprite
new x 1,272 ops/sec ±1.78% (83 runs sampled) 840 B / frame
previous x 1,001 ops/sec ±1.45% (86 runs sampled) 2520 B / frame
Fastest is new
Lowest memory is new

Moving Sprite
new x 717 ops/sec ±1.11% (86 runs sampled) 1968 B / frame
previous x 510 ops/sec ±0.95% (88 runs sampled) 4496 B / frame
Fastest is new
Lowest memory is new

Sprite with input
new x 657 ops/sec ±1.46% (86 runs sampled) 2352 B / frame
previous x 530 ops/sec ±2.00% (86 runs sampled) 4440 B / frame
Fastest is new
Lowest memory is new

Sprites removed and created
new x 167 ops/sec ±1.40% (83 runs sampled) 11272 B / frame
previous x 134 ops/sec ±0.91% (75 runs sampled) 16872 B / frame
Fastest is new
Lowest memory is new

1000 textures
new x 3.15 ops/sec ±1.50% (12 runs sampled) 440896 B / frame
previous x 3.03 ops/sec ±1.03% (12 runs sampled) 478552 B / frame
Fastest is new
Lowest memory is new

1000 Sprites
new x 0.92 ops/sec ±1.46% (7 runs sampled) 1238920 B / frame
previous x 0.65 ops/sec ±1.73% (6 runs sampled) 2451096 B / frame
Fastest is new
Lowest memory is new

1000 Pure Sprites
new x 2.32 ops/sec ±1.81% (10 runs sampled) 200112 B / frame
previous x 1.98 ops/sec ±2.04% (9 runs sampled) 563728 B / frame
Fastest is new
Lowest memory is new

200 Nested Sprites
new x 9.11 ops/sec ±0.81% (27 runs sampled) 170728 B / frame
previous x 0.76 ops/sec ±4.23% (6 runs sampled) 3934192 B / frame
Fastest is new
Lowest memory is new
```